+++ packages/analytics/analytics-metric-provider/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [11.0.2](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@11.0.1...@kong-ui-public/analytics-metric-provider@11.0.2) (2025-11-12)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
 ## [11.0.1](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@11.0.0...@kong-ui-public/analytics-metric-provider@11.0.1) (2025-11-07)
 
 **Note:** Version bump only for package @kong-ui-public/analytics-metric-provider